refactor per webcharts-wrapper-boilerplate
