+++ README.md
@@ -1,0 +1,4 @@
+# Shift Plot for Clinical Trials
+The safety-results-over-time library shows population averages for lab measures, vital signs and other related measures during the course of a clinical trial.. The library is built using [Webcharts](https://github.com/RhoInc/Webcharts) and [D3](https://www.d3js.org). 
+
+See the [wiki](https://github.com/RhoInc/safety-results-over-time/wiki) for more details.
