+++ package.json
@@ -1,16 +1,16 @@
 {
   "name": "safety-results-over-time",
-  "version": "1.1.0",
-  "description": "outlier explorer for safety explorer",
-  "main": "renderer.common.js",
+  "version": "1.2.0",
+  "description": "Chart showing population averages for lab measures, vital signs and other related measures during the course of a clinical trial.",
+  "main": ".build/safetyResultsOverTime.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "dev-watch": "cd test && node watcher.js",
-    "build": "rollup -c | babel > index.js",
-    "build-renderer": "rollup -c renderer.rollup.config.js | babel > renderer.common.js"
+    "build": "rollup -c ./scripts/rollup.wrapper.config.js | babel > ./build/safetyResultsOverTime.js",
+    "build-react": "rollup -c ./scripts/rollup.react.config.js | babel > ./build/safetyResultsOverTime-react.js"
   },
-  "author": "Rho, Inc.",
-  "license": "ISC",
+  "author": "Rho Inc",
+  "license": "MIT",
   "dependencies": {
     "d3": "^3.5.14",
     "react": "^0.14.7",
