--- renderer.rollup.config.js
@@ -1,9 +1,0 @@
-module.exports = {
-  entry: './cf-wrapper/Renderer.js',
-  format: 'cjs',
-  globals: {
-    webcharts: 'webCharts',
-    d3: 'd3',
-    react: 'React'
-  }
-};
