--- rollup.config.js
@@ -1,9 +1,0 @@
-module.exports = {
-  entry: './src/index.js',
-  format: 'iife',
-  globals: {
-    webcharts: 'webCharts',
-    d3: 'd3'
-  },
-  moduleName: 'resultsOverTime'
-};
