+++ scripts/rollup.cf-renderer.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  entry: './src/chartfoundry/bin/Renderer.js',
+  format: 'cjs',
+  globals: {
+    webcharts: 'webCharts',
+    d3: 'd3',
+    react: 'React'
+  }
+};
