+++ scripts/rollup.wrapper.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  entry: './src/wrapper.js',
+  format: 'iife',
+  globals: {
+    webcharts: 'webCharts',
+    d3: 'd3'
+  },
+  moduleName: 'safetyResultsOverTime'
+};
