--- test/index.html
@@ -1,36 +1,0 @@
-<html>
-  <head>
-    <title>Test Results Over Time</title>
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  	<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-  	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-  	<link rel="stylesheet" type="text/css" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.5.0/css/font-awesome.min.css">
-    <!-- <link rel="stylesheet" type="text/css" href="http://graphics.rhoworld.com/src/webcharts/webcharts.v1.4.0.css"> -->
-    <link rel="shortcut icon" href="chart-foundry3.ico">
-  </head>
-  <body>
-    <div id='root'>
-    </div>
-    <script src="http://d3js.org/d3.v3.min.js" charset="utf-8"></script>
-    <script src='http://graphics.rhoworld.com/src/webcharts/webcharts.v1.3.2.min.js'></script>
-    <script src="bundle.js"></script>
-    <script>
-      var div = '#root';
-      var settings = {
-        start_value: 'Bilirubin',
-        violins: true,
-        x: {
-          tickAttr: {
-            rotate: -45,
-            anchor: 'end'
-          }
-        }
-      };
-      var dataFile = 'analysis-test-results.csv';
-
-      d3.csv(dataFile, function (data){
-        var plot = resultsOverTime(div, settings).init(data);
-      });
-    </script>
-  </body>
-</html>
